requirements.txt
kiteconnect==4.2.0
python-dotenv==1.0.0
pyotp==2.9.0
requests==2.31.0
flask==3.0.0
selenium==4.15.0